+++ NEWS.md
@@ -1,6 +1,7 @@
 # mrf2d 0.2.1.9999 (Development version)
 
   * Co-ocurrence histogram computing is now exported in the `cohist()` function.
+  * Computing of conditional probabilities for individual positions are now exported in the `cp_mrf2d()` function.
   * All the conversions from/to arrays to/from vectors are now available. See the help file for `smr_stat`.
   * Additional example data included in `field1` and `hfield1`.
   * Minor documentation improvements.